+++ src/index.js
@@ -13,7 +13,7 @@ import Request from './request';
 import RequestList from './request_list';
 import { openRequestQueue } from './request_queue';
 import SettingsRotator from './settings_rotator';
-import { apifyClient, getMemoryInfo, isProduction, amIAtHome } from './utils';
+import { apifyClient, getMemoryInfo, isProduction, isAtHome } from './utils';
 import { browse, launchWebDriver } from './webdriver';
 
 /* globals module */
@@ -44,7 +44,7 @@ module.exports = {
     readyFreddy,
     getMemoryInfo,
     getApifyProxyUrl,
-    amIAtHome,
+    isAtHome,
     client: apifyClient,
 
     // Autoscaled pool