+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+xxxxxxxxxxxxxxxxxxx
+===================
+- Added `ignoreMainProcess` parameter to `AutoscaledPool`. Check documentation for more info.
+
 0.5.26 / 2018-04-27
 ===================
 - Parameters `session` and `groups` of `getApifyProxyUrl()` renamed to `apifyProxySession` and `apifyProxyGroups` to match