+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 xxxxxxxxxxxxxxxxxxx
-- Added `Apify.amIAtHome()` function that returns `true` when code is running at Apify platform and `false` otherwise
+- Added `Apify.isAtHome()` function that returns `true` when code is running at Apify platform and `false` otherwise
   (for example locally).
 
 0.5.26 / 2018-04-27